Create useReadPost hook
