+++ src/app/posts/[id]/page.tsx
@@ -2,11 +2,9 @@
 
 import clsx from 'clsx';
 import { type ReactNode, use, useEffect } from 'react';
-import { useQuery } from '@tanstack/react-query';
 
 import { PostCard, PostCardSkeleton } from '@/components';
-import { QueryKey } from '@/lib/enums';
-import { useCreateView, useSignedInUser } from '@/lib/hooks';
+import { useCreateView, useReadPost, useSignedInUser } from '@/lib/hooks';
 
 type Props = {
   params: Promise<{ id: string }>;
@@ -16,18 +14,7 @@ const Page = (props: Props): ReactNode => {
   const params = use(props.params);
   const { id } = params;
 
-  // TODO
-  const getPost = async () => {
-    const response: Response = await fetch(`/api/posts/${id}`);
-    const data = await response.json();
-
-    return data;
-  };
-
-  const { data, isLoading } = useQuery({
-    queryKey: [QueryKey.POST, id],
-    queryFn: getPost,
-  });
+  const { data, isLoading } = useReadPost(id);
 
   const { signedInUser } = useSignedInUser();
