+++ src/lib/hooks/index.ts
@@ -2,6 +2,7 @@ export { default as useSignedInUser } from './useSignedInUser';
 
 export { default as useCreatePost } from './useCreatePost';
 export { default as useDeletePost } from './useDeletePost';
+export { default as useReadPost } from './useReadPost';
 export { default as useReadPosts } from './useReadPosts';
 export { default as useUpdatePost } from './useUpdatePost';
