+++ src/lib/hooks/useReadPost.ts
@@ -1,0 +1,21 @@
+'use client';
+
+import { useQuery } from '@tanstack/react-query';
+
+import { QueryKey } from '../enums';
+
+const useReadPost = (id: string) => {
+  const getPost = async () => {
+    const response: Response = await fetch(`/api/posts/${id}`);
+    const data = await response.json();
+
+    return data;
+  };
+
+  return useQuery({
+    queryKey: [QueryKey.POST, id],
+    queryFn: getPost,
+  });
+};
+
+export default useReadPost;
